replaced int[][]s with long[][]s

+++ 04-USACO/Silver.java
@@ -2,13 +2,13 @@ import java.util.*;
 import java.io.*;
 
 public class Silver {
-  public static int solve(String filename) {
+  public static long solve(String filename) {
     try {
       Scanner Pudding = new Scanner(new File(filename)); //Pudding, the trusty file parser!
     } catch (FileNotFoundException E) {
       return -1;
     }
-    int[][] map; //2d array representing the map
+    long[][] map; //2d array representing the map
     //map details:
     //-1 means upsteppable
     //int >= 0 means steppable
@@ -19,7 +19,7 @@ public class Silver {
 
     //scan into map
     Pudding.nextLine(); //extract newLine
-    map = new int[N][M];
+    map = new long[N][M];
     for (int i = 0; i < N; i++) {
       String line = Pudding.nextLine();
       for (int k = 0; k < M; k++) {
@@ -40,14 +40,14 @@ public class Silver {
     return map[R1][C1];
   }
 
-  public static int[][] advance(int[][] array) {
-    int[][] newMap = new int[array.length][array[0].length];
+  public static long[][] advance(long[][] array) {
+    long[][] newMap = new long[array.length][array[0].length];
     for (int i = 0; i < array.length; i++) {
       for (int k = 0; k < array[i].length; k++) {
         if (array[i][k] == -1) { //if it is a tree
           newMap[i][k] = -1; //transfer tree over
         } else { //if not a tree, transfer the number of ways to step into the square
-          int sum = 0;
+          long sum = 0;
           if (valid(array, i + 1, k)) {sum += array[i + 1][k];}
           if (valid(array, i, k + 1)) {sum += array[i][k + 1];}
           if (valid(array, i - 1, k)) {sum += array[i - 1][k];}
@@ -59,13 +59,13 @@ public class Silver {
     return newMap;
   }
 
-  public static boolean valid(int[][] array, int row, int col) {
+  public static boolean valid(long[][] array, int row, int col) {
     //pray that short circuiting wont fail me
     return (row >= 0 && col >= 0 && row < array.length && col < array[row].length) && (array[row][col] != -1);
   }
 
-  public static void print2DArray(int[][] array) {
-    for (int[] row : array) {
+  public static void print2DArray(long[][] array) {
+    for (long[] row : array) {
       System.out.println(Arrays.toString(row));
       System.out.println('\n');
     }
